- Drop superflous cloning of the buffer. Gets the bench time down from 10,700,000 ns to 10,200,000 ns
- Use a custom rounding function instead of the slow-but-accurate rounding intrinsic. Gets the bench time down from 10,200,000 ns to 4,200,000 ns